requirements.txt
# Haystack Learning Project - Python Dependencies
# Core Haystack Framework and Dependencies

# Primary Framework
haystack-ai>=2.0.0
hatch>=1.9.0

# LLM Providers
openai>=1.56.1
cohere>=5.0.0
anthropic>=0.7.0

# HuggingFace Integration
transformers>=4.35.0
sentence-transformers>=2.2.0
datasets>=2.14.0
huggingface-hub>=0.18.0

# Vector Databases and Document Stores
pinecone-client>=3.0.0
weaviate-client>=4.4.0
elasticsearch>=8.8.0
redis>=5.0.0
psycopg2-binary>=2.9.0

# Data Processing and File Handling
pypdf>=4.0.0
python-docx>=1.1.0
python-pptx>=0.6.0
markdown>=3.5.0
beautifulsoup4>=4.12.0
requests>=2.31.0
aiohttp>=3.9.0

# Machine Learning and NLP
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
nltk>=3.8.0
spacy>=3.7.0

# Web Framework and API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Development and Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
black>=23.9.0
flake8>=6.1.0
mypy>=1.6.0
pre-commit>=3.5.0

# Monitoring and Observability
prometheus-client>=0.19.0
structlog>=23.1.0
python-json-logger>=2.0.0

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
click>=8.1.0
rich>=13.6.0
jinja2>=3.1.0

# Jupyter and Interactive Development
jupyter>=1.0.0
jupyterlab>=4.0.0
ipython>=8.15.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Performance and Caching
redis>=5.0.0
diskcache>=5.6.0
joblib>=1.3.0

# Security
cryptography>=41.0.0
python-jose[cryptography]>=3.3.0

# Development Dependencies (Optional)
# Uncomment if needed for specific development tasks
# tensorboard>=2.14.0
# wandb>=0.15.0
# optuna>=3.4.0
